Move notion base entity to separate module
